Update VnfLauncherRoute, removing yaml uploading Now it's a get request
